Add Page copying to Admin
Modeling Product copying, Page copying allows for a page and its content
to be cloned and worked on as a separate page.

A deprecation warning has been added to `WORKAREA.productCopyIds` as
this module was too specific to copying products. The new, generalized
`WORKAREA.copyIds` should be used instead.

WORKAREA-168